+++ R/class-args.R
@@ -3,10 +3,10 @@
 #'
 #' ReferenceClass object for managing arguments
 #'
-#' @details The `scribeArg` class sets specifications and controls for how
+#' @details The [scribeArg] class sets specifications and controls for how
 #'   command line arguments are to be parsed.  These are meant to be used in
 #'   conjunction with [scribeCommandArgs] and specifically with the [Rscript]
-#'   utility.  However, a use can define their own `scribeArg` separately.
+#'   utility.  However, a use can define their own [scribeArg] separately.
 #'
 #' @section Options:
 #'
@@ -34,8 +34,11 @@
 #' }
 #'
 #' @field aliases `[character]`\cr A vector to denote the argument's name
-#' @field action `[character]`\cr An action for resolving the argument
-#' @field default `[ANY]`\cr A default value
+#' @field action `[character]`\cr An action for resolving the argument (see
+#'   `default` for note on using another [scribeArg] object)
+#' @field default `[ANY]`\cr A default value.  This can be another [scribeArg]
+#'   object.  When that is the case, the default value and action are pass
+#'   through from the other [scribeArg] object.
 #' @field convert `[ANY]`\cr Passed to the `to` argument in [value_convert()]
 #' @field n `[integer]`\cr The length of the values
 #' @field info `[character]`\cr Additional information about the argument when
@@ -43,6 +46,8 @@
 #' @field options `[list]`\cr A named list of options (see **Options**)
 #' @field positional `[logical]`\cr Indicator if the argument is _positional_
 #'   (i.e., not preceded by a `-` or `--` command line argument)
+#' @field resolved `[logical]`\cr Has the object been resolved
+#' @field value `[ANY]`\cr The resolve value
 #'
 #' @examples
 #' # new_arg() is recommended over direct use of scribeArg$new()
@@ -71,7 +76,9 @@ scribeArg <- methods::setRefClass( # nolint: object_name_linter.
     n          = "integer",
     info       = "character",
     options    = "list",
-    positional = "logical"
+    positional = "logical",
+    resolved   = "logical",
+    value      = "ANY"
   )
 )
 
@@ -140,5 +147,15 @@ scribeArg$methods(
   get_default = function() {
     "Retrieve the default value"
     arg_get_default(.self)
+  },
+
+  get_value = function() {
+    "Retrieve the resolved value"
+    arg_get_value(.self)
+  },
+
+  is_resolved = function() {
+    "Check if object has been resolved"
+    arg_is_resolved(.self)
   }
 )